float to unsigned int for brush size and scale

+++ src/map.cpp
@@ -11,7 +11,7 @@ using namespace std;
 float Map::getWidth() const { return width_; }
 float Map::getHeight() const { return height_; }
 string Map::getName() const { return mapName_; }
-float Map::getScale() const { return scale_; }
+unsigned int Map::getScale() const { return scale_; }
 vector<Layer>& Map::getLayers() { return layers_; }
 void Map::insertLayer(const Layer& layer, int depth) { layers_.insert(layers_.begin() + depth, layer) ; }
 void Map::displayMap(sf::RenderWindow& window) {
